requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
httpx>=0.20.0
python-dotenv>=0.19.0
supabase>=1.0.0,<2.0.0 